backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
sqlalchemy==2.0.25
aiosqlite==0.19.0
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
PyJWT==2.12.0
python-jose==3.3.0
Pillow==10.1.0
requests==2.31.0
google-generativeai==0.3.0
face-recognition==1.3.5
opencv-python==4.8.1.78
dlib==19.24.2
numpy==1.24.3
